Notebooks for transfer learning

+++ rbm/core/dice.py
@@ -2,8 +2,8 @@ from keras import backend as K
 
 
 def dice_coef(y_true, y_pred):
-    y_true_f = K.flatten(y_true)
-    y_pred_f = K.flatten(y_pred)
+    y_true_f = K.cast(K.flatten(y_true),'float32')
+    y_pred_f = K.cast(K.flatten(y_pred),'float32')
     intersection = K.sum(y_true_f * y_pred_f)
     return (2.0 * intersection + K.epsilon()) / (K.sum(y_true_f) + K.sum(y_pred_f) + K.epsilon())
